backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autopep8>=2.3.2",
    "fastapi>=0.117.1",
    "httpx[socks]>=0.28.1",
    "langfuse>=3.5.1",
    "pydantic-ai>=1.0.10",
    "pydantic-ai-slim[openai]>=1.0.10",
    "aiohttp>=3.12.15",
    "docling>=2.43.0",
    "httpx[socks]>=0.28.1",
    "pathvalidate>=3.3.1",
    "pdfplumber>=0.11.7",
    "pytest>=8.4.1",
    "python-pptx>=1.0.2",
    "sqlmodel>=0.0.25",
    "dirtyjson>=1.0.8",
    "chromadb>=1.1.1",
    "dataclasses>=0.8",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"
